+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Please have a look at the [Haml Coffee Changelog](https://github.com/netzpirat/haml-coffee/blob/master/CHANGELOG.md) also.
 
+## Master
+
+- [#109][]: Make helpers API not dependant on Rails.
+
 ## 1.13.1 - August 9, 2013
 
 - [#108][]: Re-add helper path API to make it easy to integrate into Sinatra.